Free socket from originating thread.
For some reason Linux is having trouble when I try to free an object
that was allocated in one thread in a separate thread.  I added a
simple queue so that the creating thread can also do the freeing.

Note: this is not a problem on the mac, this appears to be a linux
only issue.  I tried using Valgrind to try to find any obvious memory
corruption errors, but nothing turned up.