Basic support for HTTPS MITM
This patch includes basic support for performing MITM on HTTPS
connections.  There are still some certificate issues we have to work
out with the browser, but for some sites it does seem to work.

There are still lots of print messages that I will leave in for now
while we are still debugging this feature.