Got rid of queue for freeing data and restored vanilla proxy
In this patch I fixed the memory free issue I was having with Linux
before and I got rid of that killQueue for socket freeing.  I also
rolled back the Cache.cpp file to handle generic HTTP proxy
functionality because we are going to use this proxy for a few things
and I would like to implement more protocols and more advanced HTTP
features independed of the uses for the proxy.